refactor and made functions for data and prompt generators (#9)
* refactor and made functions for data and prompt generators

* update readme

* rename and refactor set generation

* add details to functions. rename functions

+++ setlexsem/generate/generate_prompts.py
@@ -2,30 +2,26 @@
 
 import argparse
 import ast
+import itertools
 import logging
 import os
 import random
 from collections.abc import Iterable
 from itertools import product
+from typing import Dict, List, Union
 
 import pandas as pd
 from tqdm import tqdm
 
 from setlexsem.constants import PATH_PROMPTS_ROOT
+from setlexsem.generate.generate_sets import get_sampler, make_hps_set
 from setlexsem.generate.prompt import (
     PromptConfig,
     get_ground_truth,
     get_prompt,
 )
-from setlexsem.generate.sample import (
-    BasicNumberSampler,
-    BasicWordSampler,
-    DeceptiveWordSampler,
-    DecileWordSampler,
-    OverlapSampler,
-    Sampler,
-)
-from setlexsem.generate.utils_data_generation import load_generated_data
+from setlexsem.generate.sample import Sampler
+from setlexsem.generate.utils_io import load_generated_data
 from setlexsem.utils import get_prompt_file_path, read_config
 
 
@@ -42,15 +38,82 @@ def get_parser():
         required=True,
         help="Path to config file for generating prompts",
     )
+    parser.add_argument(
+        "--save-data",
+        action="store_true",
+        help="Save data to disk",
+    )
+    parser.add_argument(
+        "--overwrite", action="store_true", help="Overwrite data"
+    )
     return parser
 
 
-def create_prompt(
+def make_hps_prompt(
+    op_list=None,
+    k_shot=None,
+    prompt_type=None,
+    prompt_approach=None,
+    is_fix_shot=None,
+    config: Dict[str, List[Union[str, int]]] = {},
+):
+    """Make hyperparamters for the prompts"""
+    if config:
+        op_list = config["op_list"]
+        k_shot = config.get("k_shot")
+        prompt_type = config.get("prompt_type")
+        prompt_approach = config.get("prompt_approach")
+        is_fix_shot = config.get("is_fix_shot")
+
+    # Wrap each parameter in a list if it isn’t already, to enable Cartesian product
+    param_grid = {
+        "op_list": op_list if isinstance(op_list, list) else [op_list],
+        "k_shot": k_shot if isinstance(k_shot, list) else [k_shot],
+        "prompt_type": prompt_type
+        if isinstance(prompt_type, list)
+        else [prompt_type],
+        "prompt_approach": prompt_approach
+        if isinstance(prompt_approach, list)
+        else [prompt_approach],
+        "is_fix_shot": is_fix_shot
+        if isinstance(is_fix_shot, list)
+        else [is_fix_shot],
+    }
+
+    # Generate combinations of all parameters as dictionaries
+    keys, values = zip(*param_grid.items())
+    return (dict(zip(keys, v)) for v in product(*values))
+
+
+def get_prompt_config(
+    prompt_config: Dict[str, List[Union[str, int]]], k_shot_sampler: Sampler
+):
+    """Convert dictionary to PromptConfig class"""
+    # prepare config for prompt
+    prompt_config_ready = PromptConfig(
+        operation=prompt_config["op_list"],
+        k_shot=prompt_config["k_shot"],
+        type=prompt_config["prompt_type"],
+        approach=prompt_config["prompt_approach"],
+        sampler=k_shot_sampler,
+        is_fixed_shots=prompt_config["is_fix_shot"],
+    )
+
+    return prompt_config_ready
+
+
+def create_prompts_from_sampler(
     sampler: Sampler,
-    prompt_config: PromptConfig,
+    prompt_config: Dict[str, List[Union[str, int]]],
+    k_shot_sampler: Sampler,
     num_runs=100,
     add_roles=False,  # Claude Instant
 ):
+    """Create the prompt and the ground truth from Sampler and PromptConfig"""
+    # get prompt config
+    prompt_config_ready = get_prompt_config(prompt_config, k_shot_sampler)
+
+    results = 0
     prompt_and_ground_truth = []
     for i in tqdm(range(num_runs)):
         # create two sets from the sampler
@@ -67,26 +130,138 @@ def create_prompt(
         prompt = get_prompt(
             A,
             B,
-            prompt_config,
+            prompt_config_ready,
             add_roles=add_roles,
         )
 
-        ground_truth = get_ground_truth(prompt_config.operation, A, B)
+        ground_truth = get_ground_truth(prompt_config_ready.operation, A, B)
 
         prompt_and_ground_truth.append(
             {
                 "prompt": prompt,
                 "ground_truth": ground_truth,
-                **prompt_config.to_dict(),
+                **prompt_config_ready.to_dict(),
             }
         )
 
     return prompt_and_ground_truth
 
 
-def main(config_file):
-    LOGGER = logging.getLogger(__name__)
-    LOGGER.setLevel(level=logging.INFO)
+def create_prompts(
+    set_types=None,
+    n=None,
+    m=None,
+    item_len=None,
+    decile_group=None,
+    swap_status=None,
+    overlap_fraction=None,
+    op_list=None,
+    k_shot=None,
+    prompt_type=None,
+    prompt_approach=None,
+    is_fix_shot=None,
+    number_of_data_points=10,
+    random_seed_value=292,
+    add_roles=False,
+    data_config: Dict[str, List[Union[str, int]]] = {},
+    prompt_config: Dict[str, List[Union[str, int]]] = {},
+):
+    """Generate prompts for the given hyperparameters"""
+    # generator for set construction
+    if data_config:
+        make_hps_generator = make_hps_set(config=data_config)
+    else:
+        make_hps_generator = make_hps_set(
+            config={
+                "set_types": set_types,
+                "n": n,
+                "m": m,
+                "item_len": item_len,
+                "decile_group": decile_group,
+                "swap_status": swap_status,
+                "overlap_fraction": overlap_fraction,
+            }
+        )
+
+    # generator for prompts
+    if prompt_config:
+        make_hps_generator = make_hps_prompt(config=prompt_config)
+    else:
+        make_hps_prompt_generator = make_hps_prompt(
+            config={
+                "op_list": op_list,
+                "k_shot": k_shot,
+                "prompt_type": prompt_type,
+                "prompt_approach": prompt_approach,
+                "is_fix_shot": is_fix_shot,
+            }
+        )
+
+    make_hps_generator, make_hps_generator_copy = itertools.tee(
+        make_hps_generator
+    )
+
+    # report number of overall experiments
+    n_experiments = len(list(make_hps_prompt_generator)) * len(
+        list(make_hps_generator_copy)
+    )
+
+    # go through hyperparameters and run the experiment
+    counter_exp = 1
+    output = {}
+    for hp in make_hps_generator:
+        # generator for prompts
+        make_hps_prompt_generator = make_hps_prompt(
+            config={
+                "op_list": op_list,
+                "k_shot": k_shot,
+                "prompt_type": prompt_type,
+                "prompt_approach": prompt_approach,
+                "is_fix_shot": is_fix_shot,
+            }
+        )
+        for hp_prompt in make_hps_prompt_generator:
+            # Initilize Seed for each combination
+            random_state = random.Random(random_seed_value)
+
+            # Create Sampler()
+            try:
+                sampler = get_sampler(hp, random_state)
+            except Exception as e:
+                print(
+                    f"------> Error: Cannot create sampler. Skipping this experiment: {e}"
+                )
+                counter_exp += 1
+                continue
+
+            # create kshot sampler, before loading data
+            k_shot_sampler = sampler.create_sampler_for_k_shot()
+
+            # Create prompts
+            try:
+                prompt_and_ground_truth = create_prompts_from_sampler(
+                    sampler,
+                    prompt_config=hp_prompt,
+                    k_shot_sampler=k_shot_sampler,
+                    num_runs=number_of_data_points,
+                    add_roles=add_roles,  # Claude Instant
+                )
+            except Exception as e:
+                print(
+                    f"------> Error: Cannot create prompt. Skipping this experiment: {e}"
+                )
+                counter_exp += 1
+                continue
+
+            output[counter_exp] = prompt_and_ground_truth
+            counter_exp += 1
+
+    return output
+
+
+def main(config_file, save_data, overwrite):
+    logger = logging.getLogger(__name__)
+    logger.setLevel(level=logging.INFO)
 
     # TODO: Add this to Config
     SWAP_STATUS = False
@@ -96,6 +271,8 @@ def main(config_file):
     config = read_config(config_file)
 
     # Experiment config
+    number_of_data_points = config["N_RUN"]
+
     RANDOM_SEED_VAL = config["RANDOM_SEED_VAL"]
     OP_LIST = config["OP_LIST"]
 
@@ -118,33 +295,51 @@ def main(config_file):
     IS_FIX_SHOT = config["IS_FIX_SHOT"]
 
     # generator for prompts
-    def make_hps_prompt():
-        return product(
-            OP_LIST, K_SHOT, PROMPT_TYPE, PROMPT_APPROACH, IS_FIX_SHOT
-        )
-
+    make_hps_prompt_generator = make_hps_prompt(
+        config={
+            "op_list": OP_LIST,
+            "k_shot": K_SHOT,
+            "prompt_type": PROMPT_TYPE,
+            "prompt_approach": PROMPT_APPROACH,
+            "is_fix_shot": IS_FIX_SHOT,
+        }
+    )
     # generator for set construction
-    if DECILE_NUM[0] is not None:
-
-        def make_hps():
-            return product(
-                SET_TYPES, N, M, ITEM_LEN, DECILE_NUM, OVERLAP_FRACTION
-            )
-
-    else:
-
-        def make_hps():
-            return product(SET_TYPES, N, M, ITEM_LEN, OVERLAP_FRACTION)
+    make_hps_generator = make_hps_set(
+        config={
+            "set_types": SET_TYPES,
+            "n": N,
+            "m": M,
+            "item_len": ITEM_LEN,
+            "decile_group": DECILE_NUM,
+            "swap_status": SWAP_STATUS,
+            "overlap_fraction": OVERLAP_FRACTION,
+        }
+    )
+    make_hps_generator, make_hps_generator_copy = itertools.tee(
+        make_hps_generator
+    )
 
     # report number of overall experiments
-    n_experiments = len(list(make_hps())) * len(list(make_hps_prompt()))
-    LOGGER.info(f"Experiment will run for {n_experiments} times")
+    n_experiments = len(list(make_hps_prompt_generator)) * len(
+        list(make_hps_generator_copy)
+    )
+    logger.info(f"Creating prompts for {n_experiments} configurations...")
 
     # go through hyperparameters and run the experiment
     counter_exp = 1
-    for hp in make_hps():
-        for hp_prompt in make_hps_prompt():
-            LOGGER.info(
+    for hp in make_hps_generator:
+        make_hps_prompt_generator = make_hps_prompt(
+            config={
+                "op_list": OP_LIST,
+                "k_shot": K_SHOT,
+                "prompt_type": PROMPT_TYPE,
+                "prompt_approach": PROMPT_APPROACH,
+                "is_fix_shot": IS_FIX_SHOT,
+            }
+        )
+        for hp_prompt in make_hps_prompt_generator:
+            logger.info(
                 f"-------- EXPERIMENT #{counter_exp} out of {n_experiments}"
             )
 
@@ -152,69 +347,31 @@ def main(config_file):
             random_state = random.Random(RANDOM_SEED_VAL)
 
             # Create Sampler()
-            if hp[0] == "numbers" or "BasicNumberSampler" in hp[0]:
-                sampler = BasicNumberSampler(
-                    n=hp[1],
-                    m=hp[2],
-                    item_len=hp[3],
-                    random_state=random_state,
-                )
-            elif hp[0] == "words" or "BasicWordSampler" in hp[0]:
-                sampler = BasicWordSampler(
-                    n=hp[1],
-                    m=hp[2],
-                    item_len=hp[3],
-                    random_state=random_state,
-                )
-            elif hp[0] == "deceptive_words":
-                sampler = DeceptiveWordSampler(
-                    n=hp[1],
-                    m=hp[2],
-                    random_state=random_state,
-                    swap_set_elements=SWAP_STATUS,
-                    swap_n=hp[2] // 2,
-                )
-            elif hp[0] == "decile_words":
-                sampler = DecileWordSampler(
-                    n=hp[1], m=hp[2], item_len=hp[3], decile_num=hp[4]
-                )
-
-            # if overlapping, create OverlapSampler()
             try:
-                if "overlapping" in hp[0]:
-                    sampler = OverlapSampler(sampler, overlap_fraction=hp[4])
-            except:
-                LOGGER.error("------> Error: Skipping this experiment")
+                sampler = get_sampler(hp, random_state)
+            except Exception as e:
+                logger.error(f"------> Error: Skipping this experiment: {e}")
                 counter_exp += 1
                 continue
-            LOGGER.info(sampler)
-            # create k-shot sampler
+            logger.info(sampler)
+
+            # create kshot sampler, before loading data
             k_shot_sampler = sampler.create_sampler_for_k_shot()
 
             # load already created data
             sampler = load_generated_data(sampler, RANDOM_SEED_VAL)
 
-            # Create Prompt Config
-            prompt_config = PromptConfig(
-                operation=hp_prompt[0],
-                k_shot=hp_prompt[1],
-                type=hp_prompt[2],
-                approach=hp_prompt[3],
-                sampler=k_shot_sampler,
-                is_fixed_shots=hp_prompt[4],
-            )
-            LOGGER.info(prompt_config)
-
             # Create prompts
             try:
-                prompt_and_ground_truth = create_prompt(
+                prompt_and_ground_truth = create_prompts_from_sampler(
                     sampler,
-                    prompt_config,
-                    num_runs=100,
+                    prompt_config=hp_prompt,
+                    k_shot_sampler=k_shot_sampler,
+                    num_runs=number_of_data_points,
                     add_roles=add_roles,  # Claude Instant
                 )
             except Exception as e:
-                LOGGER.error(f"------> Error: Skipping this experiment: {e}")
+                logger.error(f"------> Error: Skipping this experiment: {e}")
                 counter_exp += 1
                 continue
 
@@ -227,13 +384,15 @@ def main(config_file):
             path_to_prompts = os.path.join(
                 PATH_PROMPTS_ROOT, folder_structure, filename
             )
-            os.makedirs(os.path.dirname(path_to_prompts), exist_ok=True)
-            # save prompts
-            pd.DataFrame(prompt_and_ground_truth).to_csv(
-                path_to_prompts, index=False
-            )
+            if save_data:
+                os.makedirs(os.path.dirname(path_to_prompts), exist_ok=True)
+                # save prompts
+                if os.path.exists(path_to_prompts) and not overwrite:
+                    pd.DataFrame(prompt_and_ground_truth).to_csv(
+                        path_to_prompts, index=False
+                    )
 
-    LOGGER.info("Done!")
+    logger.info("Done!")
 
 
 # init
@@ -242,5 +401,7 @@ if __name__ == "__main__":
     parser = get_parser()
     args = parser.parse_args()
     config_path = args.config_path
+    save_data = args.save_data
+    overwrite = args.overwrite
 
-    main(config_path)
+    main(config_path, save_data, overwrite)
